requirements.txt
opencv-python
tqdm
numpy
matplotlib
Pillow
Flask
Werkzeug
darknet
torch
torchvision
transforms
requests
aiohttp
asyncio
aiofiles
mysql-connector-python
multipledispatch
SQLAlchemy
jsonmerge
pycococreatortools
defusedxml
future
olefile
pip
wheel
cryptography
lxml
keyring
attrs
scipy
idna
multidict
pytz
Cython
setuptools
qiniu
simplejson
yarl
web.py
certifi
chardet
tifffile
scikit-image
imageio
networkx
urllib3
six
python-dateutil
decorator
pandas
PyYAML
pydot
jsonschema
pyrsistent
cycler
pygobject
Jinja2
pyparsing
kiwisolver
pycocotools
zipp